Added binary and library crates
