+++ console_sim/src/dummy_led.rs
@@ -1,13 +1,6 @@
 use std::cell::Cell;
 
-pub const PWM_MAX: u32 = 100;
-
-/// Power LED of the flashlight. [pwn] represents the duty cycle.
-pub trait Led {
-    fn set_pwm(&self, pwm: u32);
-    fn get_pwm(&self) -> u32;
-    fn modify(&self, f: &dyn Fn(u32) -> u32);
-}
+use light_control::bsp::led::Led;
 
 /// Led which resides in memory, for simulation or testing
 pub struct DummyLed {
