+++ console_sim/src/dummy_rgb.rs
@@ -1,14 +1,5 @@
 use std::cell::Cell;
-
-pub const RED: u8 = 0x4;
-pub const GREEN: u8 = 0x02;
-pub const BLUE: u8 = 0x01;
-
-/// Power LED of the flashlight. [pwn] represents the duty cycle.
-pub trait Rgb {
-    fn set_rgb(&self, rgb: u8);
-    fn get_rgb(&self) -> u8;
-}
+use light_control::bsp::rgb::Rgb;
 
 /// Led which resides in memory, for simulation or testing
 pub struct DummyRgb {
