+++ light_control/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rusty-light"
+name = "light_control"
 version = "0.1.0"
 authors = ["Yuriy Kulikov <yuriy.kulikov.87@gmail.com>"]
 edition = "2018"
@@ -7,12 +7,4 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-crossterm = "0.18.0"
-keyboard_query = "0.1.0"
-light_control = { path = "./light_control" }
-
-[[bin]]
-name = "console_sim"
-path = "console_sim/src/main.rs"
-test = false
-bench = false
+[lib]
